feat: llvm build, hardening, cleanup, support external vars.sh

+++ build.sh
@@ -5,38 +5,99 @@
 # - Mimalloc: 3.1.5
 # - OpenSSL: 3.6.0
 # - Zlib: Cloudflare Zlib @gcc.amd64
+# - Zstd: (coming soon)
 # - SQLite: 3.51.0
 # - LLVM: 21.1.2
 
-set -e -x
+# Eventual goal:
+# LLVM_PROJECTS="clang;clang-tools-extra;lldb;lld;bolt"
 
+set -e
+
+## Components.
+BUILD_ZLIB=${BUILD_ZLIB:-yes}
+BUILD_ZSTD=${BUILD_ZSTD:-no}
+BUILD_OPENSSL=${BUILD_OPENSSL:-yes}
+BUILD_LLVM=${BUILD_LLVM:-yes}
+BUILD_SQLITE=${BUILD_SQLITE:-yes}
+BUILD_CAPNP=${BUILD_CAPNP:-no}
+BUILD_STAGE2=${BUILD_STAGE2:-yes}
+
+## Settings.
+RELEASE=${RELEASE:-no}
 HARDEN=${HARDEN:-yes}
 SECURE=${SECURE:-OFF}
 GUARDED=${GUARDED:-OFF}
 CFI=${CFI:-no}
 MPK=${MPK:-no}
-CLEAN_BEFORE_BUILD=${CLEAN_BEFORE_BUILD:-no}
-MUSL_USE_MIMALLOC=${MUSL_USE_MIMALLOC:-yes}
+LLVM_PROJECTS="clang;lld;bolt"
+
+## Advanced settings.
 USE_MUSL_CROSSMAKE=${USE_MUSL_CROSSMAKE:-yes}
 USE_SCCACHE=${USE_SCCACHE:-yes}
+USE_WIDE_VECTORS=${USE_WIDE_VECTORS:-no}
+CLEAN_BEFORE_BUILD=${CLEAN_BEFORE_BUILD:-no}
+MUSL_USE_MIMALLOC=${MUSL_USE_MIMALLOC:-yes}
 MAKE_SYMLINK=${MAKE_SYMLINK:-no}
+
+MODERN_X86_64_ARCH_TARGET=x86-64-v4
+MODERN_X86_64_ARCH_TUNE=znver3
+MODERN_AARCH64_ARCH_TARGET=armv8.4-a+crypto+crc
+MODERN_AARCH64_ARCH_TUNE=neoverse-v1
+
 JOBS=`nproc`
 ARCH_FLAVOR=${ARCH_FLAVOR:-amd64}
-C_TARGET_ARCH=${C_TARGET_ARCH:-x86-64-v4}
-C_TARGET_TUNE=${C_TARGET_TUNE:-znver3}
 ROOT_DIR=$PWD
 
 MUSL_VERSION=1.2.5
 SYSROOT_PREFIX="$ROOT_DIR/$MUSL_VERSION"
 
-BUILD_ZLIB=${BUILD_ZLIB:-yes}
-BUILD_OPENSSL=${BUILD_OPENSSL:-yes}
-BUILD_SQLITE=${BUILD_SQLITE:-yes}
-BUILD_CAPNP=${BUILD_CAPNP:-no}
+OPT_CFLAGS="-O2 -ffat-lto-objects -fno-semantic-interposition"
+SECURITY_CFLAGS="-fno-plt -fstack-protector-strong -fstack-clash-protection -fomit-frame-pointer -D_FORTIFY_SOURCE=2 -Wa,--noexecstack"
+OPT_LDFLAGS=""
+SECURITY_LDFLAGS="-Wl,-z,relro,-z,now,-z,separate-code"
 
 unset CC
 unset CFLAGS
 unset LDFLAGS
+
+source ./vars.sh || echo "No vars; using defaults."
+
+echo "-----------------------------------------------"
+echo "Musl toolchain:"
+echo "BUILD_ZLIB=$BUILD_ZLIB"
+echo "BUILD_ZSTD=$BUILD_ZSTD"
+echo "BUILD_OPENSSL=$BUILD_OPENSSL"
+echo "BUILD_LLVM=$BUILD_LLVM"
+echo "BUILD_SQLITE=$BUILD_SQLITE"
+echo "BUILD_CAPNP=$BUILD_CAPNP"
+echo "BUILD_STAGE2=$BUILD_STAGE2"
+echo ""
+echo "RELEASE=$RELEASE"
+echo "HARDEN=$HARDEN"
+echo "SECURE=$SECURE"
+echo "GUARDED=$GUARDED"
+echo "CFI=$CFI"
+echo "MPK=$MPK"
+echo "LLVM_PROJECTS=$LLVM_PROJECTS"
+echo ""
+echo "USE_MUSL_CROSSMAKE=$USE_MUSL_CROSSMAKE"
+echo "USE_SCCACHE=$USE_SCCACHE"
+echo "USE_WIDE_VECTORS=$USE_WIDE_VECTORS"
+echo "CLEAN_BEFORE_BUILD=$CLEAN_BEFORE_BUILD"
+echo "MUSL_USE_MIMALLOC=$MUSL_USE_MIMALLOC"
+echo "MAKE_SYMLINK=$MAKE_SYMLINK"
+echo "ARCH_FLAVOR=$ARCH_FLAVOR"
+echo "JOBS=$JOBS"
+echo "-----------------------------------------------"
+echo "Bulding starting in 3 seconds..."
+sleep 1
+echo "2..."
+sleep 1
+echo "1..."
+sleep 1
+
+set -e -x
 
 # if musl version is empty, fail
 if [ -z "$MUSL_VERSION" ]; then
@@ -57,27 +118,40 @@ mkdir -p ./"$MUSL_VERSION/include/asm-generic"
 # Copy kernel headers (adjust paths for Ubuntu's multiarch layout)
 cp -r /usr/include/linux/* "./$MUSL_VERSION/include/linux/"
 cp -r /usr/include/asm-generic/* "./$MUSL_VERSION/include/asm-generic/"
-SECURITY_CFLAGS=""
 
 # if the ARCH_FLAVOR is amd64, we need to copy from x86_64-linux-musl
 if [ "$ARCH_FLAVOR" = "amd64" ]; then
   ARCH_FLAVOR="x86_64";
   MUSL_TARGET="x86_64-linux-musl"
   LINUX_ARCH_DIR="x86_64-linux-musl"
-  SECURITY_CFLAGS="$SECURITY_CFLAGS"
+  C_TARGET_ARCH=${C_TARGET_ARCH:-$MODERN_X86_64_ARCH_TARGET}
+  C_TARGET_TUNE=${C_TARGET_TUNE:-$MODERN_X86_64_ARCH_TUNE}
+  if [ "$USE_WIDE_VECTORS" = "yes" ]; then
+    OPT_CFLAGS="$OPT_CFLAGS -mavx512f -mavx512bw -mavx512dq -mavx512vl"
+  fi
+  SECURITY_CFLAGS="$SECURITY_CFLAGS -fcf-protection=branch"
   if [ "$CFI" = "yes" ]; then
     SECURITY_CFLAGS="$SECURITY_CFLAGS -fcf-protection=full $([ "$MPK" = "yes" ] && echo "-mpku")"
   fi
   # Specifically copy asm headers from x86_64-linux-gnu, since musl does not have them
   cp -r /usr/include/x86_64-linux-gnu/asm/* "./$MUSL_VERSION/include/asm/"
   pushd "$MUSL_VERSION/bin";
-  ln -s musl-gcc x86_64-linux-musl-gcc;
+  # create symlink if it does not exist
+  if [ ! -f x86_64-linux-musl-gcc ]; then
+    ln -s musl-gcc x86_64-linux-musl-gcc;
+  fi
   popd;
-else if [ "$ARCH_FLAVOR" = "arm64" ]; then
+elif [ "$ARCH_FLAVOR" = "arm64" ]; then
   ARCH_FLAVOR="aarch64";
   MUSL_TARGET="$ARCH_FLAVOR-linux-musl"
   LINUX_ARCH_DIR="$ARCH_FLAVOR-linux-musl"
   SECURITY_CFLAGS="$SECURITY_CFLAGS -mbranch-protection=standard"
+  C_TARGET_ARCH=${C_TARGET_ARCH:-$MODERN_AARCH64_ARCH_TARGET}
+  C_TARGET_TUNE=${C_TARGET_TUNE:-$MODERN_AARCH64_ARCH_TUNE}
+  if [ "$USE_WIDE_VECTORS" = "yes" ]; then
+    C_TARGET_ARCH="$C_TARGET_ARCH+sve"
+    OPT_CFLAGS="$OPT_CFLAGS -msve-vector-bits=256"
+  fi
   cp -r /usr/include/$ARCH_FLAVOR-linux-gnu/asm/* "./$MUSL_VERSION/include/asm/"
   pushd "$MUSL_VERSION/bin";
   ln -s musl-gcc "$ARCH_FLAVOR-linux-musl-gcc";
@@ -86,9 +160,7 @@ else
   echo "Unsupported ARCH_FLAVOR: $ARCH_FLAVOR";
   exit 1;
 fi
-fi
 
-# Verify you got what you need
 ls "./$MUSL_VERSION/include/linux/mman.h"
 
 sudo chown -R $(whoami) "./$MUSL_VERSION"
@@ -107,9 +179,9 @@ make clean && \
     --with-malloc=mallocng \
     --enable-optimize=internal,malloc,string \
   && make -j${JOBS} \
-    CFLAGS_AUTO="-march=$C_TARGET_ARCH -mtune=$C_TARGET_TUNE -O3 -ffat-lto-objects -fno-plt -fstack-protector-strong -fomit-frame-pointer -Wl,-z,relro,-z,now -Wa,--noexecstack $SECURITY_CFLAGS" \
-    CFLAGS_MEMOPS="-march=$C_TARGET_ARCH -mtune=$C_TARGET_TUNE -O3 -ffat-lto-objects $SECURITY_CFLAGS" \
-    CFLAGS_LDSO="-march=$C_TARGET_ARCH -mtune=$C_TARGET_TUNE -O3 -ffat-lto-objects $SECURITY_CFLAGS" \
+    CFLAGS_AUTO="-march=$C_TARGET_ARCH -mtune=$C_TARGET_TUNE $OPT_CFLAGS $SECURITY_CFLAGS -O3" \
+    CFLAGS_MEMOPS="-march=$C_TARGET_ARCH -mtune=$C_TARGET_TUNE $OPT_CFLAGS $SECURITY_CFLAGS -O3" \
+    CFLAGS_LDSO="-march=$C_TARGET_ARCH -mtune=$C_TARGET_TUNE $OPT_CFLAGS $SECURITY_CFLAGS -O3" \
     ADD_CFI=$CFI \
     | tee buildlog.txt;
 
@@ -133,7 +205,9 @@ if [ "$USE_MUSL_CROSSMAKE" = "yes" ]; then
   cp -fv config.mak musl-cross-make/config.mak
   pushd musl-cross-make;
 
-  make clean || echo "Nothing to clean.";
+  if [ "$CLEAN_BEFORE_BUILD" = "yes" ]; then
+    make clean || echo "Nothing to clean.";
+  fi
 
   make -j${JOBS} \
     MUSL_ARCH="$ARCH_FLAVOR" \
@@ -153,11 +227,11 @@ fi
 
 export CC="$SYSROOT_PREFIX/bin/musl-gcc"
 
-export CFLAGS="-I$SYSROOT_PREFIX/include -I/usr/include/$LINUX_ARCH_DIR"
-export LDFLAGS="-L$SYSROOT_PREFIX/lib"
+export CFLAGS="-I$SYSROOT_PREFIX/include -I/usr/include/$LINUX_ARCH_DIR -march=$C_TARGET_ARCH -mtune=$C_TARGET_TUNE $OPT_CFLAGS $SECURITY_CFLAGS $CFLAGS"
+export LDFLAGS="-L$SYSROOT_PREFIX/lib $OPT_LDFLAGS $SECURITY_LDFLAGS"
 
-export CFLAGS="-march=$C_TARGET_ARCH -mtune=$C_TARGET_TUNE -O2 -ffat-lto-objects -fstack-protector-strong -Wl,-z,relro,-z,now -Wa,--noexecstack -D_FORTIFY_SOURCE=2 $SECURITY_CFLAGS $CFLAGS"
-export LDFLAGS="-ffat-lto-objects $LDFLAGS"
+# Add LDFLAGS here if needed.
+# export LDFLAGS="$LDFLAGS"
 
 ## Build mimalloc
 echo "------- Building mimalloc...";
@@ -186,37 +260,42 @@ popd;
 echo "Mimalloc done."
 
 ## Build musl (phase 2 + mimalloc)
-echo "------- Building musl (phase 2)...";
-pushd musl;
+if [ "$BUILD_STAGE2" = "yes" ]; then
+  echo "------- Building musl (phase 2)...";
+  pushd musl;
 
-mkdir -p mimalloc/objs;
-rm -fv buildlog.txt mimalloc/objs/*;
-rm -f config.mak
-make distclean || true
+  mkdir -p mimalloc/objs;
+  rm -fv buildlog.txt mimalloc/objs/*;
+  rm -f config.mak
+  make distclean || true
 
-if [ "$MUSL_USE_MIMALLOC" = "yes" ]; then
-  cp -fv ../mimalloc/out/release/mimalloc*.o ./mimalloc/objs/;
-else
-  echo "Not using mimalloc in musl build.";
+  if [ "$MUSL_USE_MIMALLOC" = "yes" ]; then
+    cp -fv ../mimalloc/out/release/mimalloc*.o ./mimalloc/objs/;
+  else
+    echo "Not using mimalloc in musl build.";
+  fi
+
+  make clean && \
+    ./configure \
+      --prefix="$SYSROOT_PREFIX" \
+      --enable-optimize=internal,malloc,string \
+    && make -j${JOBS} \
+      CFLAGS_AUTO="-march=$C_TARGET_ARCH -mtune=$C_TARGET_TUNE $OPT_CFLAGS $SECURITY_CFLAGS -O3" \
+      CFLAGS_MEMOPS="-march=$C_TARGET_ARCH -mtune=$C_TARGET_TUNE $OPT_CFLAGS $SECURITY_CFLAGS -O3" \
+      CFLAGS_LDSO="-march=$C_TARGET_ARCH -mtune=$C_TARGET_TUNE $OPT_CFLAGS $SECURITY_CFLAGS -O3" \
+      USE_MIMALLOC=$MUSL_USE_MIMALLOC \
+      ADD_CFI=$CFI \
+      | tee buildlog.txt;
+
+  make install;
+
+  popd;
 fi
 
-make clean && \
-  ./configure \
-    --prefix="$SYSROOT_PREFIX" \
-    --enable-optimize=internal,malloc,string \
-  && make -j${JOBS} \
-    CFLAGS_AUTO="-march=$C_TARGET_ARCH -mtune=$C_TARGET_TUNE -O3 -ffat-lto-objects -fno-plt -fstack-protector-strong -fomit-frame-pointer -Wl,-z,relro,-z,now $SECURITY_CFLAGS" \
-    CFLAGS_MEMOPS="-march=$C_TARGET_ARCH -mtune=$C_TARGET_TUNE -O3 -ffat-lto-objects $SECURITY_CFLAGS" \
-    CFLAGS_LDSO="-march=$C_TARGET_ARCH -mtune=$C_TARGET_TUNE -O3 -ffat-lto-objects $SECURITY_CFLAGS" \
-    USE_MIMALLOC=$MUSL_USE_MIMALLOC \
-    ADD_CFI=$CFI \
-    | tee buildlog.txt;
-
-make install;
-
-popd;
-
 ## Mount musl sysroot
+
+CC_NO_PREFIX="$MUSL_GCC"
+CXX_NO_PREFIX="$MUSL_GCC"
 
 if [ "$USE_SCCACHE" = "yes" ]; then
   export CC="`which sccache` $MUSL_GCC"
@@ -226,8 +305,8 @@ else
   export CXX="$MUSL_GCC"
 fi
 
-export CFLAGS="$CFLAGS -flto=auto"
-export LDFLAGS="$LDFLAGS -flto=auto -static -Wl,-z,relro,-z,now,-z,noexecstack"
+export CFLAGS="$CFLAGS -O3 -flto=auto"
+export LDFLAGS="$LDFLAGS -flto=auto -static"
 
 ## Build zlib
 
@@ -236,9 +315,11 @@ if [ "$BUILD_ZLIB" != "yes" ]; then
 else
   echo "------- Building zlib...";
   pushd zlib;
-  git checkout .
-  git clean -xdf
-  make clean || echo "Nothing to clean.";
+  if [ "$CLEAN_BEFORE_BUILD" = "yes" ]; then
+    git checkout .
+    git clean -xdf
+    make clean || echo "Nothing to clean.";
+  fi
   ./configure \
     --const \
     --64 \
@@ -249,6 +330,59 @@ else
   popd;
 fi
 
+### Build zstd (coming soon)
+
+### Build LLVM
+if [ "$BUILD_LLVM" != "yes" ]; then
+  echo "Skipping LLVM build.";
+else
+  echo "------- Building LLVM...";
+  pushd llvm;
+
+  if [ "$CLEAN_BEFORE_BUILD" = "yes" ]; then
+    echo "Cleaning previous LLVM build ..."
+    rm -fr build
+    git checkout .
+    git clean -xdf
+  fi
+
+  mkdir -p build;
+  popd;
+  pushd llvm/build;
+  cmake ../llvm \
+    -DCMAKE_BUILD_TYPE=Release \
+    -DCMAKE_INSTALL_PREFIX="$SYSROOT_PREFIX" \
+    -DCMAKE_C_COMPILER="$SYSROOT_PREFIX/bin/${MUSL_TARGET}-gcc" \
+    -DCMAKE_CXX_COMPILER="$SYSROOT_PREFIX/bin/${MUSL_TARGET}-g++" \
+    -DCMAKE_C_COMPILER_LAUNCHER=$(which sccache) \
+    -DCMAKE_CXX_COMPILER_LAUNCHER=$(which sccache) \
+    -DLLVM_ENABLE_PROJECTS="$LLVM_PROJECTS" \
+    -DLLVM_TOOL_BOLT_BUILD=TRUE \
+    -DLLVM_TOOL_CLANG_BUILD=TRUE \
+    -DLLVM_BUILD_32_BITS=OFF \
+    -DLLVM_BUILD_BENCHMARKS=OFF \
+    -DLLVM_BUILD_DOCS=OFF \
+    -DLLVM_BUILD_EXAMPLES=OFF \
+    -DLLVM_ENABLE_EH=ON \
+    -DLLVM_ENABLE_PIC=OFF \
+    -DLLVM_BUILD_STATIC=ON \
+    -DLLVM_ENABLE_ZLIB=ON \
+    -DLLVM_ENABLE_ZSTD=ON \
+    -DLLVM_ENABLE_RPMALLOC=ON \
+    -DLLVM_TARGETS_TO_BUILD="X86;AArch64" \
+    -DLLVM_DEFAULT_TARGET_TRIPLE="$MUSL_TARGET" \
+    -DLIBCLANG_BUILD_STATIC=ON \
+    -DCLANG_LINK_CLANG_DYLIB=OFF \
+    -DCMAKE_EXE_LINKER_FLAGS="-static -L$SYSROOT_PREFIX/$LINUX_ARCH_DIR/lib -latomic" \
+    -DCMAKE_SHARED_LINKER_FLAGS="-L$SYSROOT_PREFIX/$LINUX_ARCH_DIR/lib -latomic" \
+    -DCMAKE_MODULE_LINKER_FLAGS="-L$SYSROOT_PREFIX/$LINUX_ARCH_DIR/lib -latomic" \
+    -DLLVM_INTEGRATED_CRT_ALLOC=OFF \
+    -DLLVM_ENABLE_RTTI=ON \
+    && make -j `nproc` \
+    && make install;
+  popd;
+fi
+
 ### Build openssl
 
 if [ "$BUILD_OPENSSL" != "yes" ]; then
@@ -256,13 +390,15 @@ if [ "$BUILD_OPENSSL" != "yes" ]; then
 else
   echo "------- Building openssl...";
   pushd openssl;
-  git checkout .
-  git clean -xdf
-  make clean || echo "No clean step";
+  if [ "$CLEAN_BEFORE_BUILD" = "yes" ]; then
+    git checkout .
+    git clean -xdf
+    make clean || echo "No clean step";
+  fi
   rm -fv buildlog.txt;
   export CFLAGS="$CFLAGS -fPIE -fPIC"
 
-  if [ "$ARCH_FLAVOR" = "amd64" ]; then
+  if [ "$ARCH_FLAVOR" = "x86_64" ]; then
     OPENSSL_TARGET="linux-x86_64"
   else
     OPENSSL_TARGET="linux-aarch64"
@@ -274,10 +410,10 @@ else
       no-tests \
       no-external-tests \
       no-comp \
-      no-asm \
       no-afalgeng \
       enable-ec_nistp_64_gcc_128 \
       enable-tls1_3 \
+      enable-asm \
       threads \
       -static \
       --prefix="$SYSROOT_PREFIX" \
@@ -298,9 +434,11 @@ if [ "$BUILD_SQLITE" != "yes" ]; then
 else
   echo "------- Building sqlite...";
   pushd sqlite;
-  git checkout .
-  git clean -xdf
-  make clean || echo "No clean step";
+  if [ "$CLEAN_BEFORE_BUILD" = "yes" ]; then
+    git checkout .
+    git clean -xdf
+    make clean || echo "No clean step";
+  fi
   rm -fv buildlog.txt;
 
   ./configure \
@@ -323,6 +461,10 @@ if [ "$BUILD_CAPNP" != "yes" ]; then
 else
   echo "------- Building capnp...";
 
+  if [ "$CLEAN_BEFORE_BUILD" = "yes" ]; then
+    make clean || echo "Nothing to clean.";
+  fi
+
   cd capnp/c++;
   autoreconf -i;
   ./configure \
@@ -343,7 +485,7 @@ file "$SYSROOT_PREFIX/bin/musl-gcc" || exit 2
 file "$SYSROOT_PREFIX/lib/libc.a" || exit 2
 file "$SYSROOT_PREFIX/lib/libz.a" || exit 3
 
-if [ "$ARCH_FLAVOR" = "amd64" ]; then
+if [ "$ARCH_FLAVOR" = "x86_64" ]; then
   OPENSSL_LIB_DIR="lib64"
 else
   OPENSSL_LIB_DIR="lib"
@@ -371,11 +513,17 @@ echo ""
 echo "Components:"
 echo "  mimalloc:   3.1.5"
 echo "  libc:       $(file "$SYSROOT_PREFIX/lib/libc.a")"
+if [ "$BUILD_LLVM" = "yes" ]; then
+  echo "  llvm:       21.1.2 $(file "$SYSROOT_PREFIX/bin/clang")"
+fi
 if [ "$BUILD_ZLIB" = "yes" ]; then
   echo "  zlib:       (cloudflare@gcc.amd64) $(file "$SYSROOT_PREFIX/lib/libz.a")"
 fi
+if [ "$BUILD_ZSTD" = "yes" ]; then
+  echo "  zstd:       (coming soon) $(file "$SYSROOT_PREFIX/lib/libzstd.a")"
+fi
 if [ "$BUILD_OPENSSL" = "yes" ]; then
-  echo "  openssl:    3.6.0 $(file "$SYSROOT_PREFIX/lib64/libssl.a")"
+  echo "  openssl:    3.6.0 $(file "$SYSROOT_PREFIX/$OPENSSL_LIB_DIR/libssl.a")"
 fi
 if [ "$BUILD_SQLITE" = "yes" ]; then
   echo "  sqlite:     3.51.0 $(file "$SYSROOT_PREFIX/lib/libsqlite3.a")"
@@ -392,6 +540,7 @@ echo "  CFI:            $CFI"
 echo "  Mimalloc:       yes"
 echo "  Musl+Mimalloc:  $MUSL_USE_MIMALLOC"
 echo "  Musl-CrossMake: $USE_MUSL_CROSSMAKE"
+echo "  LLVM Projects:  $LLVM_PROJECTS"
 echo "-----------------------------------------------"
 
 if [ "$MAKE_SYMLINK" != "yes" ]; then
